:lipstick: Design: TeamDtailPageLeftMenuFix
